refactor: move module styles to /styles folder
