+++ src/views/containers/SearchRepo/styles/SearchHeader.style.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+import { AppColors } from '../../../../constants/AppColors';
+import { FontSize } from '../../../../constants/FontSize';
+
+export default StyleSheet.create({
+  container: {
+    height: 60,
+    backgroundColor: AppColors.PRIMARY
+  },
+  searchInput: {
+    fontSize: FontSize.NORMAL,
+    color: AppColors.WHITE,
+    backgroundColor: AppColors.DARK_GREY,
+    margin: 8,
+    borderRadius: 8,
+    paddingHorizontal: 16
+  }
+});
